+++ src/data/v1/bars.rs
@@ -10,9 +10,10 @@ use serde::Deserialize;
 use serde::Serialize;
 use serde_urlencoded::to_string as to_query;
 
+use time_util::optional_system_time_to_rfc3339;
+use time_util::system_time_from_secs;
+
 use crate::data::DATA_BASE_URL;
-use crate::time_util::optional_system_time_to_rfc3339;
-use crate::time_util::system_time_from_secs;
 use crate::Str;